Refactor new Folder view a little bit
