+++ src/webapp/src/components/Library/lists/albums/song-list/index.js
@@ -9,7 +9,7 @@ import {
   Typography,
 } from '@mui/material';
 
-import request from '../../../utils/request';
+import request from '../../../../../utils/request';
 
 import SongListHeader from './song-list-header';
 import SongListHeadline from './song-list-headline';
